Split AST behavior collection
