+++ src/typechecker/mod.rs
@@ -16,6 +16,7 @@ mod behavior_ref_metadata;
 mod closures;
 mod declaration_collection;
 mod declaration_collection_ast;
+mod declaration_collection_ast_behaviors;
 mod declaration_collection_ast_callables;
 mod declaration_collection_resolver_semantic_tasks;
 mod declaration_collection_resolver_tasks;
