+++ tests/docs_truth/repo_hygiene/mod.rs
@@ -20,6 +20,7 @@ mod source_truth;
 mod typechecker_aggregate_constructors;
 mod typechecker_behavior_impls;
 mod typechecker_call_validation;
+mod typechecker_declaration_collection;
 mod typechecker_imports;
 mod typechecker_program_checking;
 mod typechecker_resolver_validation;
